silence compiler warnings

+++ src/pmo_c.cpp
@@ -86,7 +86,7 @@ void polympo_getMPVelArray(mpmesh mpMeshIn, int size, double* array) {
   auto mpVel = MPs->getData<polyMpmTest::MPF_Vel>();
   auto mpVelCopy = polyMpmTest::DoubleVec2DView("copyOfMPVel",size);
     
-  PMT_ALWAYS_ASSERT(MPs->getCount()*vec2d_nEntries==mpVelCopy.size());
+  PMT_ALWAYS_ASSERT(static_cast<size_t>(MPs->getCount()*vec2d_nEntries)==mpVelCopy.size());
 
   auto copyVel = PS_LAMBDA(const int& elm, const int& mp, const int& mask){
     if(mask) {
